Refactor settings migration logic
