+++ frontend/src/hooks/mutation/use-save-settings.ts
@@ -1,8 +1,10 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { ApiSettings, Settings } from "#/services/settings";
+import {
+  ApiSettings,
+  LATEST_SETTINGS_VERSION,
+  Settings,
+} from "#/services/settings";
 import OpenHands from "#/api/open-hands";
-import { updateSettingsVersion } from "#/utils/settings-utils";
-import { useAuth } from "#/context/auth-context";
 import { useSettingsUpToDate } from "#/context/settings-up-to-date-context";
 
 const saveSettingsMutationFn = async (settings: Partial<Settings>) => {
@@ -21,15 +23,19 @@ const saveSettingsMutationFn = async (settings: Partial<Settings>) => {
 
 export const useSaveSettings = () => {
   const queryClient = useQueryClient();
-  const { logout } = useAuth();
-  const { setIsUpToDate } = useSettingsUpToDate();
+  const { isUpToDate, setIsUpToDate } = useSettingsUpToDate();
 
   return useMutation({
     mutationFn: saveSettingsMutationFn,
     onSuccess: async () => {
       await queryClient.invalidateQueries({ queryKey: ["settings"] });
-      updateSettingsVersion(logout);
-      setIsUpToDate(true);
+      if (!isUpToDate) {
+        localStorage.setItem(
+          "SETTINGS_VERSION",
+          LATEST_SETTINGS_VERSION.toString(),
+        );
+        setIsUpToDate(true);
+      }
     },
   });
 };
