+++ frontend/src/utils/settings-utils.ts
@@ -1,9 +1,4 @@
-import {
-  settingsAreUpToDate,
-  maybeMigrateSettings,
-  LATEST_SETTINGS_VERSION,
-  Settings,
-} from "#/services/settings";
+import { Settings } from "#/services/settings";
 
 const extractBasicFormData = (formData: FormData) => {
   const provider = formData.get("llm-provider")?.toString();
@@ -78,18 +73,4 @@ const saveSettingsView = (view: "basic" | "advanced") => {
   );
 };
 
-/**
- * Updates the settings version in local storage if the current settings are not up to date.
- * If the settings are outdated, it attempts to migrate them before updating the version.
- */
-const updateSettingsVersion = async (logout: () => void) => {
-  if (!settingsAreUpToDate()) {
-    await maybeMigrateSettings(logout);
-    localStorage.setItem(
-      "SETTINGS_VERSION",
-      LATEST_SETTINGS_VERSION.toString(),
-    );
-  }
-};
-
-export { extractSettings, saveSettingsView, updateSettingsVersion };
+export { extractSettings, saveSettingsView };
